Move Segment Model to preferences package.
